requirements.txt
mysql-connector-python==8.0.22
pymysql==0.10.1
